Gera valores de anova para BIBD

Calcula valores utilizados para uma tabela de ANOVA de um delineamento em blocos incompletos balanceados

anova_bibd(dados, x, y, bloco)

Arguments

dados
Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos

Value

Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.

Examples

bloco_incompleto(4, 6, 2)$dados %>%
 anova_bibd("Trat", "resultado", "bloco")

#> $a
#> [1] 4
#> 
#> $b
#> [1] 6
#> 
#> $N
#> [1] 12
#> 
#> $r
#> [1] 3
#> 
#> $k
#> [1] 2
#> 
#> $lambda
#> [1] 1
#> 
#> $y..
#> [1] 30
#> 
#> $yi.
#>  1  2  3  4 
#>  3  6  9 12 
#> 
#> $y.j
#> B1 B2 B3 B4 B5 B6 
#>  4  5  5  7  6  3 
#> 
#> $mu
#> [1] 2.5
#> 
#> $tau
#>   1   2   3   4 
#> 0.5 3.5 6.5 9.5 
#> 
#> $beta
#>  B1  B2  B3  B4  B5  B6 
#> 1.5 2.5 2.5 4.5 3.5 0.5 
#> 
#> $sqt
#> [1] 15
#> 
#> $sqtrat
#> [1] 15
#> 
#> $sqbloco
#> [1] 5
#> 
#> $sqe
#> [1] -5
#> 
#> $nij
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    0    0    0    1
#> [2,]    0    0    1    0    1    1
#> [3,]    1    0    1    1    0    0
#> [4,]    0    1    0    1    1    0
#> 
#> $Qi
#>   [,1]
#> 1   -3
#> 2   -1
#> 3    1
#> 4    3
#> 
#> $sqtrat_aj
#> [1] 160
#> 
#> $qmtrat_aj
#> [1] 53.33333
#> 
#> $qmbloco
#> [1] 1
#> 
#> $qme
#> [1] -1.666667
#> 
#> $f0
#> [1] -32
#> 
#> $pvalor
#> [1] 1
#> 
#> $estimados
#> # A tibble: 1 x 5
#>   `$\\hat{\\beta_i}$` `$\\hat{\\beta}$` `$\\hat{\\mu_i}$` `$\\hat{\\mu}$`
#>                 <dbl>             <dbl>             <dbl>           <dbl>
#> 1                   5               2.5               7.5             2.5
#> # … with 1 more variable: $\hat{\tau}$ <dbl>
#> 
#> $modelo
#> $$\color{red}{y_{ij} = 2.5 +
#> \tau_i + \beta_j + e_{ij}}$$
#>