Gera valores de anova para Quadrados Latinos

Calcula valores utilizados para uma tabela de ANOVA de um delineamento em quadrados latinos

anova_lsqd(dados, x, y, linha, coluna)

Arguments

dados
Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

linha
String com o nome da coluna do data frame que representa a linha do quadrado latino

coluna
String com o nome da coluna do data frame que representa a coluna do quadrado latino

Value

Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.

Examples

quadrados_latinos(tibble(y = rnorm(16)), 1:4) %>%
 anova_lsqd("tratamento", "y", "linha", "coluna")

#> $N
#> [1] 16
#> 
#> $p
#> [1] 4
#> 
#> $y...
#> [1] 0.2576366
#> 
#> $y.j.
#>          1          2          3          4 
#>  0.1833499 -2.7213431  3.6946093 -0.8989795 
#> 
#> $yi..
#>          1          2          3          4 
#> -3.0209483  0.2195031  0.9395602  2.1195217 
#> 
#> $y..k
#>           1           2           3           4 
#> -1.20832831  0.05066497  0.35945626  1.05584368 
#> 
#> $mu
#> [1] 0.01610229
#> 
#> $tau
#>          1          2          3          4 
#>  0.1672476 -2.7374454  3.6785070 -0.9150818 
#> 
#> $alpha
#>          1          2          3          4 
#> -3.0370506  0.2034008  0.9234579  2.1034194 
#> 
#> $beta
#>           1           2           3           4 
#> -1.22443059  0.03456268  0.34335397  1.03974139 
#> 
#> $sqt
#> [1] 13.88572
#> 
#> $sqtrat
#> [1] 5.470258
#> 
#> $sqlinha
#> [1] 3.633215
#> 
#> $sqcol
#> [1] 0.6725112
#> 
#> $sqe
#> [1] 4.109733
#> 
#> $qmtrat
#> [1] 1.823419
#> 
#> $qmlinha
#> [1] 1.211072
#> 
#> $qmcol
#> [1] 0.2241704
#> 
#> $qme
#> [1] 0.6849556
#> 
#> $f0
#> [1] 2.662099
#> 
#> $pvalor
#> [1] 0.1419738
#> 
#> $estimados
#> # A tibble: 1 x 7
#>   `$\\hat{\\alpha_i}$` `$\\hat{\\alpha}$` `$\\hat{\\beta_i}$` `$\\hat{\\beta}$`
#>                  <dbl>              <dbl>               <dbl>             <dbl>
#> 1               0.0644             0.0483              0.0644            0.0483
#> # … with 3 more variables: $\hat{\mu}$ <dbl>, $\hat{\tau_i}$ <dbl>,
#> #   $\hat{\tau}$ <dbl>
#> 
#> $modelo
#> $$\color{red}{y_{ij} = 0.0161022878215088 +
#> \alpha_i +\tau_k + \beta_k + e_{ij}}$$
#>