Gera valores de anova para RBCD

Calcula valores utilizados para uma tabela de ANOVA de um delineamento em blocos completamente casualizados

anova_rbcd(dados, x, y, bloco)

Arguments

dados
Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos

Value

Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.

Examples

anova_rbcd(mtcars, "gear", "hp", "carb")

#> $a
#> [1] 3
#> 
#> $b
#> [1] 6
#> 
#> $N
#> [1] 32
#> 
#> $y..
#> [1] 4694
#> 
#> $yi.
#>    3    4    5 
#> 2642 1074  978 
#> 
#> $y.j
#>    1    2    3    4    6    8 
#>  602 1172  540 1870  175  335 
#> 
#> $mu
#> [1] 146.6875
#> 
#> $tau
#>         3         4         5 
#> 2495.3125  927.3125  831.3125 
#> 
#> $beta
#>         1         2         3         4         6         8 
#>  455.3125 1025.3125  393.3125 1723.3125   28.3125  188.3125 
#> 
#> $sqt
#> [1] 145726.9
#> 
#> $sqtrat
#> [1] 826469.5
#> 
#> $sqbloco
#> [1] 1200562
#> 
#> $sqe
#> [1] -1881304
#> 
#> $qmtrat
#> [1] 413234.8
#> 
#> $qmbloco
#> [1] 240112.3
#> 
#> $qme
#> [1] -188130.4
#> 
#> $f0
#> [1] -2.196533
#> 
#> $pvalor
#> [1] 1
#> 
#> $estimados
#> # A tibble: 1 x 5
#>   `$\\hat{\\beta_i}$` `$\\hat{\\beta}$` `$\\hat{\\mu_i}$` `$\\hat{\\mu}$`
#>                 <dbl>             <dbl>             <dbl>           <dbl>
#> 1                782.              636.             1565.            147.
#> # … with 1 more variable: $\hat{\tau}$ <dbl>
#> 
#> $modelo
#> $$\color{red}{y_{ij} = 146.6875 +
#> \tau_i + \beta_j + e_{ij}}$$
#>