Simula delineamento de blocos incompletos

Simula um delineamento em blocos incompletos a partir de numeros de tratamentos, blocos e repetições.

bloco_incompleto(ntrat, nbloco, nrep = NULL)

Arguments

ntrat
Numero de tratamentos a serem simulados

nbloco
Numero de blocos a serem simulados

nrep
Numero de repeticoes a serem simuladas, caso haja, padrao = NULL

Value

Objeto de tipo lista com os elementos:

dados Dados simulados em formato longo

dados_matriz Dados simulados em formato largo, como matiz

kable Texto em LaTeX da tabela pré-formatada

matriz Matriz dos dados simulados

Examples

bloco_incompleto(8, 14, 3)$dados_matriz %>% kable_()

#> 
#> 
#> | Trat | B01 | B02 | B03 | B04 | B05 | B06 | B07 | B08 | B09 | B10 | B11 | B12 | B13 | B14 |
#> |:----:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
#> |  1   |  -  |  -  |  -  |  1  |  -  |  -  |  1  |  -  |  -  |  -  |  1  |  1  |  -  |  1  |
#> |  2   |  2  |  -  |  2  |  2  |  2  |  -  |  2  |  2  |  -  |  -  |  2  |  -  |  2  |  -  |
#> |  3   |  -  |  -  |  3  |  -  |  3  |  3  |  3  |  3  |  -  |  -  |  -  |  3  |  -  |  -  |
#> |  4   |  -  |  4  |  -  |  -  |  -  |  4  |  -  |  -  |  -  |  4  |  -  |  -  |  -  |  -  |
#> |  5   |  -  |  5  |  -  |  -  |  5  |  -  |  -  |  -  |  -  |  -  |  -  |  -  |  5  |  -  |
#> |  6   |  6  |  6  |  -  |  6  |  -  |  -  |  -  |  -  |  6  |  6  |  -  |  6  |  -  |  6  |
#> |  7   |  -  |  -  |  7  |  -  |  -  |  -  |  -  |  -  |  7  |  -  |  7  |  -  |  7  |  -  |
#> |  8   |  8  |  -  |  -  |  -  |  -  |  8  |  -  |  8  |  8  |  8  |  -  |  -  |  -  |  8  |

bloco_incompleto(4, 6, 2)

#> $dados
#> # A tibble: 24 x 3
#>     Trat resultado bloco
#>    <int>     <dbl> <chr>
#>  1     1        NA B1   
#>  2     2        NA B1   
#>  3     3         3 B1   
#>  4     4         4 B1   
#>  5     1         1 B2   
#>  6     2        NA B2   
#>  7     3        NA B2   
#>  8     4         4 B2   
#>  9     1        NA B3   
#> 10     2         2 B3   
#> # … with 14 more rows
#> 
#> $dados_matriz
#> # A tibble: 4 x 7
#>    Trat B1    B2    B3    B4    B5    B6   
#>   <int> <chr> <chr> <chr> <chr> <chr> <chr>
#> 1     1 -     1     -     1     -     1    
#> 2     2 -     -     2     -     2     2    
#> 3     3 3     -     3     3     -     -    
#> 4     4 4     4     -     -     4     -    
#> 
#> $kable
#> [1] "\n\\begin{tabular}{c|cc}\n\\hline\nTrat & resultado & bloco\\\\\n1 & NA & B1\\\\\n2 & NA & B1\\\\\n3 & 3 & B1\\\\\n4 & 4 & B1\\\\\n1 & 1 & B2\\\\\n2 & NA & B2\\\\\n3 & NA & B2\\\\\n4 & 4 & B2\\\\\n1 & NA & B3\\\\\n2 & 2 & B3\\\\\n3 & 3 & B3\\\\\n4 & NA & B3\\\\\n1 & 1 & B4\\\\\n2 & NA & B4\\\\\n3 & 3 & B4\\\\\n4 & NA & B4\\\\\n1 & NA & B5\\\\\n2 & 2 & B5\\\\\n3 & NA & B5\\\\\n4 & 4 & B5\\\\\n1 & 1 & B6\\\\\n2 & 2 & B6\\\\\n3 & NA & B6\\\\\n4 & NA & B6\\\\\n\\end{tabular}"
#> 
#> $matriz
#>    1  2  3  4  5  6
#> 1 NA  1 NA  1 NA  1
#> 2 NA NA  2 NA  2  2
#> 3  3 NA  3  3 NA NA
#> 4  4  4 NA NA  4 NA
#>