Comparacoes multiplas para BIBD

Análises múltiplas para ANOVA de delineamentos em blocos incompletos balanceados

comparacoes_bib(x, y, bloco = NULL, dados, alpha = 0.05, group = FALSE)

Arguments

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos, manter NULL caso não hajam blocos

dados
Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

Value

Objeto do tipo comparacao_bib, que contém:

dados Dados alimentados a função

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

waller Dados das comparações de waller

snk Dados das comparações de snk

Examples

bloco_incompleto(4, 6, 2)$dados ->.;
 comparacoes_bib("Trat", "resultado", "bloco", .)

#> Warning: attempting model selection on an essentially perfect fit is nonsense

#> $variaveis
#> $variaveis$tratamento
#> [1] "Trat"
#> 
#> $variaveis$bloco
#> [1] "bloco"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ Trat + bloco
#> <environment: 0x55f60ec9bb58>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 Trat bloco Residuals
#> Sum of Squares    15     0         0
#> Deg. of Freedom    3     5         3
#> 
#> Residual standard error: 1.786179e-16
#> Estimated effects may be unbalanced
#> 
#> $modelos$Modelo_aj
#> Single term deletions
#> 
#> Model:
#> resultado ~ Trat + bloco
#>        Df Sum of Sq RSS     AIC    F value  Pr(>F)    
#> <none>                0 -868.91                       
#> Trat    3        10  10    9.81 1.0448e+32 < 2e-16 ***
#> bloco   5         0   0 -844.82 9.6767e+00 0.04541 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> $modelos$lsd
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>   Mean Efficiency           CV
#>    2.5  0.6666667 2.773336e-14
#> 
#> $comparison
#>       Difference pvalue sig.
#> 1 - 2         -1      0  ***
#> 1 - 3         -2      0  ***
#> 1 - 4         -3      0  ***
#> 2 - 3         -1      0  ***
#> 2 - 4         -2      0  ***
#> 3 - 4         -1      0  ***
#> 
#> $means
#>   y mean.adj           SE r std Min Max Q25 Q50 Q75
#> 1 1        1 4.693894e-16 3   0   1   1   1   1   1
#> 2 2        2 4.693894e-16 3   0   2   2   2   2   2
#> 3 3        3 4.693894e-16 3   0   3   3   3   3   3
#> 4 4        4 4.693894e-16 3   0   4   4   4   4   4
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>   Mean Efficiency           CV
#>    2.5  0.6666667 2.773336e-14
#> 
#> $comparison
#>       Difference pvalue sig.
#> 1 - 2         -1      0  ***
#> 1 - 3         -2      0  ***
#> 1 - 4         -3      0  ***
#> 2 - 3         -1      0  ***
#> 2 - 4         -2      0  ***
#> 3 - 4         -1      0  ***
#> 
#> $means
#>   y mean.adj           SE r std Min Max Q25 Q50 Q75
#> 1 1        1 4.693894e-16 3   0   1   1   1   1   1
#> 2 2        2 4.693894e-16 3   0   2   2   2   2   2
#> 3 3        3 4.693894e-16 3   0   3   3   3   3   3
#> 4 4        4 4.693894e-16 3   0   4   4   4   4   4
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>   Mean Efficiency           CV
#>    2.5  0.6666667 2.773336e-14
#> 
#> $comparison
#>       Difference pvalue sig.
#> 1 - 2         -1      0  ***
#> 1 - 3         -2      0  ***
#> 1 - 4         -3      0  ***
#> 2 - 3         -1      0  ***
#> 2 - 4         -2      0  ***
#> 3 - 4         -1      0  ***
#> 
#> $means
#>   y mean.adj           SE r std Min Max Q25 Q50 Q75
#> 1 1        1 4.693894e-16 3   0   1   1   1   1   1
#> 2 2        2 4.693894e-16 3   0   2   2   2   2   2
#> 3 3        3 4.693894e-16 3   0   3   3   3   3   3
#> 4 4        4 4.693894e-16 3   0   4   4   4   4   4
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$waller
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>   Mean Efficiency           CV
#>    2.5  0.6666667 2.773336e-14
#> 
#> $comparison
#>       Difference significant
#> 1 - 2         -1        TRUE
#> 1 - 3         -2        TRUE
#> 1 - 4         -3        TRUE
#> 2 - 3         -1        TRUE
#> 2 - 4         -2        TRUE
#> 3 - 4         -1        TRUE
#> 
#> $means
#>   y mean.adj           SE r std Min Max Q25 Q50 Q75
#> 1 1        1 4.693894e-16 3   0   1   1   1   1   1
#> 2 2        2 4.693894e-16 3   0   2   2   2   2   2
#> 3 3        3 4.693894e-16 3   0   3   3   3   3   3
#> 4 4        4 4.693894e-16 3   0   4   4   4   4   4
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$snk
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>   Mean Efficiency           CV
#>    2.5  0.6666667 2.773336e-14
#> 
#> $comparison
#>       Difference pvalue sig.
#> 1 - 2         -1      0  ***
#> 1 - 3         -2      0  ***
#> 1 - 4         -3      0  ***
#> 2 - 3         -1      0  ***
#> 2 - 4         -2      0  ***
#> 3 - 4         -1      0  ***
#> 
#> $means
#>   y mean.adj           SE r std Min Max Q25 Q50 Q75
#> 1 1        1 4.693894e-16 3   0   1   1   1   1   1
#> 2 2        2 4.693894e-16 3   0   2   2   2   2   2
#> 3 3        3 4.693894e-16 3   0   3   3   3   3   3
#> 4 4        4 4.693894e-16 3   0   4   4   4   4   4
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 12 x 3
#>    Trat  resultado bloco
#>    <fct>     <dbl> <fct>
#>  1 2             2 B1   
#>  2 4             4 B1   
#>  3 1             1 B2   
#>  4 2             2 B2   
#>  5 2             2 B3   
#>  6 3             3 B3   
#>  7 1             1 B4   
#>  8 3             3 B4   
#>  9 1             1 B5   
#> 10 4             4 B5   
#> 11 3             3 B6   
#> 12 4             4 B6   
#> 
#> $lsd
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2                  -1       0
#> 2      1 - 3                  -2       0
#> 3      1 - 4                  -3       0
#> 4      2 - 3                  -1       0
#> 5      2 - 4                  -2       0
#> 6      3 - 4                  -1       0
#> 
#> $tukey
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2                  -1       0
#> 2      1 - 3                  -2       0
#> 3      1 - 4                  -3       0
#> 4      2 - 3                  -1       0
#> 5      2 - 4                  -2       0
#> 6      3 - 4                  -1       0
#> 
#> $duncan
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2                  -1       0
#> 2      1 - 3                  -2       0
#> 3      1 - 4                  -3       0
#> 4      2 - 3                  -1       0
#> 5      2 - 4                  -2       0
#> 6      3 - 4                  -1       0
#> 
#> $waller
#>   Comparação Diferença de Médias Significante
#> 1      1 - 2                  -1         TRUE
#> 2      1 - 3                  -2         TRUE
#> 3      1 - 4                  -3         TRUE
#> 4      2 - 3                  -1         TRUE
#> 5      2 - 4                  -2         TRUE
#> 6      3 - 4                  -1         TRUE
#> 
#> $snk
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2                  -1       0
#> 2      1 - 3                  -2       0
#> 3      1 - 4                  -3       0
#> 4      2 - 3                  -1       0
#> 5      2 - 4                  -2       0
#> 6      3 - 4                  -1       0
#> 
#> attr(,"class")
#> [1] "comparacao_bib"