Comparacoes (Quadrados Latinos)

Análises múltiplas para ANOVA de delineamentos em quadrados latinos

comparacoes_lsqd(dados, x, y, linha, coluna, alpha = 0.05, group = FALSE)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

bloco
String com o nome da coluna dos blocos, manter NULL caso não hajam blocos

Value

Objeto do tipo comparacao_lsqd, que contém:

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

dados Dados alimentados a função

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

Examples

quadrados_latinos(4)$dados %>%
  comparacoes_lsqd("tratamento", "resultado", "linha", "coluna")

#> $variaveis
#> $variaveis$tratamento
#> [1] "tratamento"
#> 
#> $variaveis$linha
#> [1] "linha"
#> 
#> $variaveis$coluna
#> [1] "coluna"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ tratamento + linha + coluna
#> <environment: 0x558f124e0b78>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 tratamento    linha   coluna Residuals
#> Sum of Squares    1.162380 2.264954 0.966973  7.811957
#> Deg. of Freedom          3        3        3         6
#> 
#> Residual standard error: 1.141049
#> Estimated effects may be unbalanced
#> 
#> $modelos$fisher
#> $statistics
#>    MSerror Df        Mean       CV  t.value      LSD
#>   1.301993  6 0.001851515 61627.85 2.446912 1.974275
#> 
#> $parameters
#>         test p.ajusted     name.t ntr alpha
#>   Fisher-LSD      none tratamento   4  0.05
#> 
#> $means
#>     resultado       std r        LCL      UCL        Min       Max        Q25
#> 1  0.45397753 0.5584482 4 -0.9420456 1.850001 -0.2753020 1.0439389  0.2101769
#> 2 -0.17509292 1.6002839 4 -1.5711161 1.220930 -2.2181411 1.4172813 -1.0119290
#> 3 -0.22794917 0.5716705 4 -1.6239723 1.168074 -0.8671011 0.3766451 -0.6153486
#> 4 -0.04352938 0.6940566 4 -1.4395525 1.352494 -0.5953490 0.9721363 -0.3758670
#>           Q50        Q75
#> 1  0.52363662 0.76743724
#> 2  0.05024411 0.88708020
#> 3 -0.21067036 0.17672908
#> 4 -0.27545242 0.05688515
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2  0.62907045 0.4652         -1.345204 2.603345
#> 1 - 3  0.68192670 0.4304         -1.292348 2.656202
#> 1 - 4  0.49750691 0.5601         -1.476768 2.471782
#> 2 - 3  0.05285626 0.9499         -1.921419 2.027131
#> 2 - 4 -0.13156354 0.8758         -2.105838 1.842711
#> 3 - 4 -0.18441980 0.8268         -2.158695 1.789855
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$bonferroni
#> $statistics
#>    MSerror Df        Mean       CV  t.value      MSD
#>   1.301993  6 0.001851515 61627.85 3.862991 3.116829
#> 
#> $parameters
#>         test  p.ajusted     name.t ntr alpha
#>   Fisher-LSD bonferroni tratamento   4  0.05
#> 
#> $means
#>     resultado       std r        LCL      UCL        Min       Max        Q25
#> 1  0.45397753 0.5584482 4 -0.9420456 1.850001 -0.2753020 1.0439389  0.2101769
#> 2 -0.17509292 1.6002839 4 -1.5711161 1.220930 -2.2181411 1.4172813 -1.0119290
#> 3 -0.22794917 0.5716705 4 -1.6239723 1.168074 -0.8671011 0.3766451 -0.6153486
#> 4 -0.04352938 0.6940566 4 -1.4395525 1.352494 -0.5953490 0.9721363 -0.3758670
#>           Q50        Q75
#> 1  0.52363662 0.76743724
#> 2  0.05024411 0.88708020
#> 3 -0.21067036 0.17672908
#> 4 -0.27545242 0.05688515
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2  0.62907045      1         -2.487758 3.745899
#> 1 - 3  0.68192670      1         -2.434902 3.798756
#> 1 - 4  0.49750691      1         -2.619322 3.614336
#> 2 - 3  0.05285626      1         -3.063973 3.169685
#> 2 - 4 -0.13156354      1         -3.248392 2.985265
#> 3 - 4 -0.18441980      1         -3.301249 2.932409
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $statistics
#>    MSerror Df        Mean       CV      MSD
#>   1.301993  6 0.001851515 61627.85 2.793059
#> 
#> $parameters
#>    test     name.t ntr StudentizedRange alpha
#>   Tukey tratamento   4         4.895599  0.05
#> 
#> $means
#>     resultado       std r        Min       Max        Q25         Q50
#> 1  0.45397753 0.5584482 4 -0.2753020 1.0439389  0.2101769  0.52363662
#> 2 -0.17509292 1.6002839 4 -2.2181411 1.4172813 -1.0119290  0.05024411
#> 3 -0.22794917 0.5716705 4 -0.8671011 0.3766451 -0.6153486 -0.21067036
#> 4 -0.04352938 0.6940566 4 -0.5953490 0.9721363 -0.3758670 -0.27545242
#>          Q75
#> 1 0.76743724
#> 2 0.88708020
#> 3 0.17672908
#> 4 0.05688515
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2  0.62907045 0.8611         -2.163989 3.422130
#> 1 - 3  0.68192670 0.8318         -2.111133 3.474986
#> 1 - 4  0.49750691 0.9231         -2.295552 3.290566
#> 2 - 3  0.05285626 0.9999         -2.740203 2.845916
#> 2 - 4 -0.13156354 0.9983         -2.924623 2.661496
#> 3 - 4 -0.18441980 0.9953         -2.977479 2.608639
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $statistics
#>    MSerror Df        Mean       CV
#>   1.301993  6 0.001851515 61627.85
#> 
#> $parameters
#>     test     name.t ntr alpha
#>   Duncan tratamento   4  0.05
#> 
#> $duncan
#>      Table CriticalRange
#> 2 3.460456      1.974275
#> 3 3.586498      2.046185
#> 4 3.648934      2.081806
#> 
#> $means
#>     resultado       std r        Min       Max        Q25         Q50
#> 1  0.45397753 0.5584482 4 -0.2753020 1.0439389  0.2101769  0.52363662
#> 2 -0.17509292 1.6002839 4 -2.2181411 1.4172813 -1.0119290  0.05024411
#> 3 -0.22794917 0.5716705 4 -0.8671011 0.3766451 -0.6153486 -0.21067036
#> 4 -0.04352938 0.6940566 4 -0.5953490 0.9721363 -0.3758670 -0.27545242
#>          Q75
#> 1 0.76743724
#> 2 0.88708020
#> 3 0.17672908
#> 4 0.05688515
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2  0.62907045 0.4786         -1.417114 2.675255
#> 1 - 3  0.68192670 0.4480         -1.399880 2.763733
#> 1 - 4  0.49750691 0.5601         -1.476768 2.471782
#> 2 - 3  0.05285626 0.9499         -1.921418 2.027131
#> 2 - 4 -0.13156354 0.8758         -2.105838 1.842711
#> 3 - 4 -0.18441980 0.8319         -2.230605 1.861765
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 16 x 4
#>    resultado tratamento coluna linha
#>        <dbl> <fct>      <fct>  <fct>
#>  1    -0.531 3          1      1    
#>  2     0.710 2          1      2    
#>  3    -0.248 4          1      3    
#>  4     0.675 1          1      4    
#>  5    -0.595 4          2      1    
#>  6     0.110 3          2      2    
#>  7     0.372 1          2      3    
#>  8    -0.610 2          2      4    
#>  9     1.04  1          3      1    
#> 10    -0.303 4          3      2    
#> 11     1.42  2          3      3    
#> 12    -0.867 3          3      4    
#> 13    -2.22  2          4      1    
#> 14    -0.275 1          4      2    
#> 15     0.377 3          4      3    
#> 16     0.972 4          4      4    
#> 
#> $fisher
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2          0.62907045       -1.345204        2.603345  0.4652
#> 2      1 - 3          0.68192670       -1.292348        2.656202  0.4304
#> 3      1 - 4          0.49750691       -1.476768        2.471782  0.5601
#> 4      2 - 3          0.05285626       -1.921419        2.027131  0.9499
#> 5      2 - 4         -0.13156354       -2.105838        1.842711  0.8758
#> 6      3 - 4         -0.18441980       -2.158695        1.789855  0.8268
#> 
#> $bonferroni
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2          0.62907045       -2.487758        3.745899       1
#> 2      1 - 3          0.68192670       -2.434902        3.798756       1
#> 3      1 - 4          0.49750691       -2.619322        3.614336       1
#> 4      2 - 3          0.05285626       -3.063973        3.169685       1
#> 5      2 - 4         -0.13156354       -3.248392        2.985265       1
#> 6      3 - 4         -0.18441980       -3.301249        2.932409       1
#> 
#> $tukey
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2          0.62907045       -2.163989        3.422130  0.8611
#> 2      1 - 3          0.68192670       -2.111133        3.474986  0.8318
#> 3      1 - 4          0.49750691       -2.295552        3.290566  0.9231
#> 4      2 - 3          0.05285626       -2.740203        2.845916  0.9999
#> 5      2 - 4         -0.13156354       -2.924623        2.661496  0.9983
#> 6      3 - 4         -0.18441980       -2.977479        2.608639  0.9953
#> 
#> $duncan
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2          0.62907045       -1.417114        2.675255  0.4786
#> 2      1 - 3          0.68192670       -1.399880        2.763733  0.4480
#> 3      1 - 4          0.49750691       -1.476768        2.471782  0.5601
#> 4      2 - 3          0.05285626       -1.921418        2.027131  0.9499
#> 5      2 - 4         -0.13156354       -2.105838        1.842711  0.8758
#> 6      3 - 4         -0.18441980       -2.230605        1.861765  0.8319
#> 
#> attr(,"class")
#> [1] "comparacao_lsqd"