Estimando valores ausentes de BIBD

Estima os valores ausentes de um delineamento incompleto balanceado

estimando_ausentes(dados, x, y, bloco)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, com valores faltantes, e blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos

Value

Data frame com os valores estimados

Examples

bloco_incompleto(4, 6, 2)$dados %>%
  estimando_ausentes("Trat", "resultado", "bloco")

#> # A tibble: 24 x 3
#>    Trat  resultado bloco   
#>    <fct>     <dbl> <chr>   
#>  1 1        0.0884 bloco B1
#>  2 2       -1.57   bloco B1
#>  3 3       -0.0798 bloco B1
#>  4 4       -1.11   bloco B1
#>  5 1        0.230  bloco B2
#>  6 2        1.73   bloco B2
#>  7 3        1.06   bloco B2
#>  8 4       -0.358  bloco B2
#>  9 1       -1.44   bloco B3
#> 10 2       -0.223  bloco B3
#> # … with 14 more rows