Simula delineamento de quadrados latinos

Simula um delineamento em quadrados latinos a partir de um banco de dados

quadrados_latinos(n, nrep = 1, fun = rnorm)

Arguments

n
Número de linhas (ou colunas)

fun
Função que gere valores aleatórios cujo primeiro argumento é a quantidade de valores a serem gerados, Padrão = rnorm()

Value

Banco de dados com colunas dos tratamentos, linhas e colunas

Examples

quadrados_latinos(4)

#> $dados_resumidos
#> # A tibble: 16 x 4
#>    resultado tratamento coluna linha
#>        <dbl> <chr>       <int> <int>
#>  1    0.333  [3]             1     1
#>  2    1.48   [2]             1     2
#>  3   -0.0111 [4]             1     3
#>  4   -0.678  [1]             1     4
#>  5   -0.359  [4]             2     1
#>  6    0.602  [3]             2     2
#>  7   -0.750  [1]             2     3
#>  8   -0.397  [2]             2     4
#>  9   -1.93   [1]             3     1
#> 10   -2.21   [4]             3     2
#> 11    1.43   [2]             3     3
#> 12   -0.240  [3]             3     4
#> 13   -0.0294 [2]             4     1
#> 14   -2.81   [1]             4     2
#> 15   -0.694  [3]             4     3
#> 16    0.132  [4]             4     4
#> 
#> $dados
#> # A tibble: 16 x 5
#>    resultado tratamento coluna linha replica
#>        <dbl> <chr>       <int> <int>   <int>
#>  1    0.333  3               1     1       1
#>  2    1.48   2               1     2       1
#>  3   -0.0111 4               1     3       1
#>  4   -0.678  1               1     4       1
#>  5   -0.359  4               2     1       1
#>  6    0.602  3               2     2       1
#>  7   -0.750  1               2     3       1
#>  8   -0.397  2               2     4       1
#>  9   -1.93   1               3     1       1
#> 10   -2.21   4               3     2       1
#> 11    1.43   2               3     3       1
#> 12   -0.240  3               3     4       1
#> 13   -0.0294 2               4     1       1
#> 14   -2.81   1               4     2       1
#> 15   -0.694  3               4     3       1
#> 16    0.132  4               4     4       1
#> 
#> $dados_matriz
#> # A tibble: 4 x 5
#>   linha `1`   `2`   `3`   `4`  
#>   <int> <chr> <chr> <chr> <chr>
#> 1     1 [3]   [4]   [1]   [2]  
#> 2     2 [2]   [3]   [4]   [1]  
#> 3     3 [4]   [1]   [2]   [3]  
#> 4     4 [1]   [2]   [3]   [4]  
#> 
#> $matriz
#>   1     2     3     4    
#> 1 "[3]" "[4]" "[1]" "[2]"
#> 2 "[2]" "[3]" "[4]" "[1]"
#> 3 "[4]" "[1]" "[2]" "[3]"
#> 4 "[1]" "[2]" "[3]" "[4]"
#>