Regressão polinomial

Regressões polinomiais de delineamentos

reg_polinomial(
  dados,
  x,
  y,
  bloco = NULL,
  grau = NULL,
  alpha = 0.05,
  uniroot_interval = NULL
)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, linhas, colunas e, caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos, caso haja

grau
grau maximo a ser considerado no modelo

alpha
Nivel de significancia a ser utilizado, padrao 0.05

uniroot_interval
Vetor com dois valores indicando o intervalo para o calculo da raiz da derivada para o calculo do PMET.

Value

Lista que contém:

formulas Lista com as formulas de todos os possíveis modelos de regressão até o grau definido.

formula_strLista com as strings das formulas de todos os modelos de ANOVA de todos os possíveis modelos de regressão até o grau definido.

modelos Lista com os modelos de ANOVA de todos os possíveis modelos de regressão até o grau definido.

falta_ajuste Lista com as faltas de ajuste de todos os possíveis modelos de regressão até o grau definido.

modelo Modelo com o melhor ajuste.

grau Grau do modelo com melhor ajuste.

plot Gráfico dos dados com o modelo sobreposto.

coeficiente Coeficientes do modelo.

fun Função do modelo.

derivada Derivada da função do modelo.

raiz Valor da raiz da função do modelo.

pmet Ponto de máxima eficiência técnica.

Examples

bloco_casualizado(4, 1, 5)$dados %>%
  reg_polinomial("Trat", "resultado")

#> Warning: Não foi encontrado nenhum modelo significativo com alpha = 0.05

#> Error in if (is.na(pmet) & is.na(raiz) & idmodelo == 1) {    pmet <- fun(raiz)}: argument is of length zero