Gera valores de anova para Quadrados Latinos

Calcula valores utilizados para uma tabela de ANOVA de um delineamento em quadrados latinos

anova_lsqd(dados, x, y, linha, coluna, replica = "")

Arguments

dados
Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

linha
String com o nome da coluna do data frame que representa a linha do quadrado latino

coluna
String com o nome da coluna do data frame que representa a coluna do quadrado latino

replica
String com o nome da coluna do data frame que representa as replicas

Value

Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.

Examples

quadrados_latinos(4)$dados %>%
 anova_lsqd("tratamento", "resultado", "linha", "coluna")

#> $stat
#> $stat$N
#> [1] 16
#> 
#> $stat$p
#> [1] 4
#> 
#> $stat$n
#> [1] 1
#> 
#> $stat$y...
#> [1] 3.851342
#> 
#> $stat$yi..
#>          1          2          3          4 
#>  0.6839904  0.8583160 -1.1515700  3.4606059 
#> 
#> $stat$y.j.
#>           1           2           3           4 
#> -0.57939977  2.84800118  1.55413837  0.02860251 
#> 
#> $stat$y..k
#>          1          2          3          4 
#>  2.9790861 -2.2112836  0.1847734  2.8987663 
#> 
#> $stat$mu
#> [1] 0.2407089
#> 
#> $stat$tau
#> [1] -0.8201087  2.6072923  1.3134295 -0.2121064
#> 
#> $stat$alpha
#> [1]  0.4432815  0.6176071 -1.3922789  3.2198970
#> 
#> $stat$beta
#> [1]  2.73837726 -2.45199253 -0.05593547  2.65805746
#> 
#> $stat$tau_
#> [1] 0.7221267
#> 
#> $stat$alpha_
#> [1] 0.7221267
#> 
#> $stat$beta_
#> [1] 0.7221267
#> 
#> $stat$tau_i
#> [1] 0.9628356
#> 
#> $stat$alpha_i
#> [1] 0.9628356
#> 
#> $stat$beta_i
#> [1] 0.9628356
#> 
#> 
#> $anova_df
#>   Fonte.de.variação Graus.de.liberdade Quadrado.Médio        F0   p.valor
#> 1        tratamento                  3      0.5962308 0.5240617 0.6815724
#> 2             linha                  3      0.8998539 0.7909335 0.5417102
#> 3            coluna                  3      1.5411257 1.3545843 0.3429141
#> 4             Erros                  9      1.1377112 1.1377112       NaN
#> 5             Total                 15            NaN       NaN       NaN
#> 
#> $anova_list
#> $anova_list$sqt
#> [1] 15.9379
#> 
#> $anova_list$sqtrat
#> [1] 1.788692
#> 
#> $anova_list$sqlinha
#> [1] 2.699562
#> 
#> $anova_list$sqcoluna
#> [1] 4.623377
#> 
#> $anova_list$sqe
#> [1] 6.826267
#> 
#> $anova_list$qmtrat
#> [1] 0.5962308
#> 
#> $anova_list$qmlinha
#> [1] 0.8998539
#> 
#> $anova_list$qmcoluna
#> [1] 1.541126
#> 
#> $anova_list$f0
#> [1] 0.5240617
#> 
#> $anova_list$pvalor
#> [1] 0.6815724
#> 
#> 
#> $estimados
#>   X..hat..mu.. X..hat..tau.. X..hat..alpha.. X..hat..beta.. X..hat..tau._i.
#> 1    0.2407089     0.7221267       0.7221267      0.7221267       0.9628356
#>   X..hat..alpha._i. X..hat..beta._i.
#> 1         0.9628356        0.9628356
#> 
#> $pvalor
#> [1] 0.6815724
#>