Comparacoes multiplas para BIBD

Análises múltiplas para ANOVA de delineamentos em blocos incompletos balanceados

comparacoes_bib(x, y, bloco = NULL, dados, alpha = 0.05, group = FALSE)

Arguments

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

bloco
String com o nome da coluna dos blocos, manter NULL caso não hajam blocos

dados
Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

Value

Objeto do tipo comparacao_bib, que contém:

dados Dados alimentados a função

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

waller Dados das comparações de waller

snk Dados das comparações de snk

Examples

bloco_incompleto(4, 6, 2)$dados ->.;
 comparacoes_bib("Trat", "resultado", "bloco", .)

#> $variaveis
#> $variaveis$tratamento
#> [1] "Trat"
#> 
#> $variaveis$bloco
#> [1] "bloco"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ Trat + bloco
#> <environment: 0x5583588e87d8>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                     Trat    bloco Residuals
#> Sum of Squares  0.386510 5.589670  3.231282
#> Deg. of Freedom        3        5         3
#> 
#> Residual standard error: 1.037831
#> Estimated effects may be unbalanced
#> 
#> $modelos$Modelo_aj
#> Single term deletions
#> 
#> Model:
#> resultado ~ Trat + bloco
#>        Df Sum of Sq    RSS     AIC F value Pr(>F)
#> <none>              3.2313  2.2557               
#> Trat    3    0.6920 3.9232 -1.4158  0.2141 0.8811
#> bloco   5    5.5897 8.8210  4.3067  1.0379 0.5217
#> 
#> $modelos$lsd
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>         Mean Efficiency       CV
#>   -0.1792375  0.6666667 -579.026
#> 
#> $comparison
#>        Difference pvalue sig.
#> 1 - 2 -0.70045100 0.5480     
#> 1 - 3 -0.21381740 0.8500     
#> 1 - 4 -0.64780643 0.5768     
#> 2 - 3  0.48663360 0.6712     
#> 2 - 4  0.05264457 0.9628     
#> 3 - 4 -0.43398903 0.7040     
#> 
#> $means
#>             y    mean.adj        SE r       std        Min        Max
#> 1 -0.18871143 -0.56975617 0.7026152 3 0.7483085 -0.6315460  0.6752700
#> 2 -0.43420779  0.13069484 0.7026152 3 0.2725916 -0.6448584 -0.1263339
#> 3  0.07292534 -0.35593876 0.7026152 3 0.3312124 -0.2753020  0.3839876
#> 4 -0.16695594  0.07805026 0.7026152 3 1.9148114 -2.2947942  1.4172813
#>          Q25        Q50         Q75
#> 1 -0.6207022 -0.6098583  0.03270586
#> 2 -0.5881448 -0.5314311 -0.32888249
#> 3 -0.0826058  0.1100904  0.24703902
#> 4 -0.9590745  0.3766451  0.89696317
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>         Mean Efficiency       CV
#>   -0.1792375  0.6666667 -579.026
#> 
#> $comparison
#>        Difference pvalue sig.
#> 1 - 2 -0.70045100 0.9005     
#> 1 - 3 -0.21381740 0.9963     
#> 1 - 4 -0.64780643 0.9180     
#> 2 - 3  0.48663360 0.9611     
#> 2 - 4  0.05264457 0.9999     
#> 3 - 4 -0.43398903 0.9716     
#> 
#> $means
#>             y    mean.adj        SE r       std        Min        Max
#> 1 -0.18871143 -0.56975617 0.7026152 3 0.7483085 -0.6315460  0.6752700
#> 2 -0.43420779  0.13069484 0.7026152 3 0.2725916 -0.6448584 -0.1263339
#> 3  0.07292534 -0.35593876 0.7026152 3 0.3312124 -0.2753020  0.3839876
#> 4 -0.16695594  0.07805026 0.7026152 3 1.9148114 -2.2947942  1.4172813
#>          Q25        Q50         Q75
#> 1 -0.6207022 -0.6098583  0.03270586
#> 2 -0.5881448 -0.5314311 -0.32888249
#> 3 -0.0826058  0.1100904  0.24703902
#> 4 -0.9590745  0.3766451  0.89696317
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>         Mean Efficiency       CV
#>   -0.1792375  0.6666667 -579.026
#> 
#> $comparison
#>        Difference pvalue sig.
#> 1 - 2 -0.70045100 0.5366     
#> 1 - 3 -0.21381740 0.8500     
#> 1 - 4 -0.64780643 0.5737     
#> 2 - 3  0.48663360 0.6684     
#> 2 - 4  0.05264457 0.9627     
#> 3 - 4 -0.43398903 0.7040     
#> 
#> $means
#>             y    mean.adj        SE r       std        Min        Max
#> 1 -0.18871143 -0.56975617 0.7026152 3 0.7483085 -0.6315460  0.6752700
#> 2 -0.43420779  0.13069484 0.7026152 3 0.2725916 -0.6448584 -0.1263339
#> 3  0.07292534 -0.35593876 0.7026152 3 0.3312124 -0.2753020  0.3839876
#> 4 -0.16695594  0.07805026 0.7026152 3 1.9148114 -2.2947942  1.4172813
#>          Q25        Q50         Q75
#> 1 -0.6207022 -0.6098583  0.03270586
#> 2 -0.5881448 -0.5314311 -0.32888249
#> 3 -0.0826058  0.1100904  0.24703902
#> 4 -0.9590745  0.3766451  0.89696317
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$waller
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>         Mean Efficiency       CV
#>   -0.1792375  0.6666667 -579.026
#> 
#> $comparison
#>        Difference significant
#> 1 - 2 -0.70045100       FALSE
#> 1 - 3 -0.21381740       FALSE
#> 1 - 4 -0.64780643       FALSE
#> 2 - 3  0.48663360       FALSE
#> 2 - 4  0.05264457       FALSE
#> 3 - 4 -0.43398903       FALSE
#> 
#> $means
#>             y    mean.adj        SE r       std        Min        Max
#> 1 -0.18871143 -0.56975617 0.7026152 3 0.7483085 -0.6315460  0.6752700
#> 2 -0.43420779  0.13069484 0.7026152 3 0.2725916 -0.6448584 -0.1263339
#> 3  0.07292534 -0.35593876 0.7026152 3 0.3312124 -0.2753020  0.3839876
#> 4 -0.16695594  0.07805026 0.7026152 3 1.9148114 -2.2947942  1.4172813
#>          Q25        Q50         Q75
#> 1 -0.6207022 -0.6098583  0.03270586
#> 2 -0.5881448 -0.5314311 -0.32888249
#> 3 -0.0826058  0.1100904  0.24703902
#> 4 -0.9590745  0.3766451  0.89696317
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$snk
#> $parameters
#>   lambda treatmeans blockSize blocks r alpha test
#>        1          4         2      6 3  0.05  BIB
#> 
#> $statistics
#>         Mean Efficiency       CV
#>   -0.1792375  0.6666667 -579.026
#> 
#> $comparison
#>        Difference pvalue sig.
#> 1 - 2 -0.70045100 0.9005     
#> 1 - 3 -0.21381740 0.8500     
#> 1 - 4 -0.64780643 0.8183     
#> 2 - 3  0.48663360 0.8900     
#> 2 - 4  0.05264457 0.9627     
#> 3 - 4 -0.43398903 0.7040     
#> 
#> $means
#>             y    mean.adj        SE r       std        Min        Max
#> 1 -0.18871143 -0.56975617 0.7026152 3 0.7483085 -0.6315460  0.6752700
#> 2 -0.43420779  0.13069484 0.7026152 3 0.2725916 -0.6448584 -0.1263339
#> 3  0.07292534 -0.35593876 0.7026152 3 0.3312124 -0.2753020  0.3839876
#> 4 -0.16695594  0.07805026 0.7026152 3 1.9148114 -2.2947942  1.4172813
#>          Q25        Q50         Q75
#> 1 -0.6207022 -0.6098583  0.03270586
#> 2 -0.5881448 -0.5314311 -0.32888249
#> 3 -0.0826058  0.1100904  0.24703902
#> 4 -0.9590745  0.3766451  0.89696317
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 12 x 3
#>    Trat  resultado bloco
#>    <fct>     <dbl> <fct>
#>  1 2        -0.126 B1   
#>  2 3         0.384 B1   
#>  3 2        -0.645 B2   
#>  4 4        -2.29  B2   
#>  5 1        -0.632 B3   
#>  6 2        -0.531 B3   
#>  7 1         0.675 B4   
#>  8 3         0.110 B4   
#>  9 1        -0.610 B5   
#> 10 4         1.42  B5   
#> 11 3        -0.275 B6   
#> 12 4         0.377 B6   
#> 
#> $lsd
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2         -0.70045100  0.5480
#> 2      1 - 3         -0.21381740  0.8500
#> 3      1 - 4         -0.64780643  0.5768
#> 4      2 - 3          0.48663360  0.6712
#> 5      2 - 4          0.05264457  0.9628
#> 6      3 - 4         -0.43398903  0.7040
#> 
#> $tukey
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2         -0.70045100  0.9005
#> 2      1 - 3         -0.21381740  0.9963
#> 3      1 - 4         -0.64780643  0.9180
#> 4      2 - 3          0.48663360  0.9611
#> 5      2 - 4          0.05264457  0.9999
#> 6      3 - 4         -0.43398903  0.9716
#> 
#> $duncan
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2         -0.70045100  0.5366
#> 2      1 - 3         -0.21381740  0.8500
#> 3      1 - 4         -0.64780643  0.5737
#> 4      2 - 3          0.48663360  0.6684
#> 5      2 - 4          0.05264457  0.9627
#> 6      3 - 4         -0.43398903  0.7040
#> 
#> $waller
#>   Comparação Diferença de Médias Significante
#> 1      1 - 2         -0.70045100        FALSE
#> 2      1 - 3         -0.21381740        FALSE
#> 3      1 - 4         -0.64780643        FALSE
#> 4      2 - 3          0.48663360        FALSE
#> 5      2 - 4          0.05264457        FALSE
#> 6      3 - 4         -0.43398903        FALSE
#> 
#> $snk
#>   Comparação Diferença de Médias P-valor
#> 1      1 - 2         -0.70045100  0.9005
#> 2      1 - 3         -0.21381740  0.8500
#> 3      1 - 4         -0.64780643  0.8183
#> 4      2 - 3          0.48663360  0.8900
#> 5      2 - 4          0.05264457  0.9627
#> 6      3 - 4         -0.43398903  0.7040
#> 
#> attr(,"class")
#> [1] "comparacao_bib"