Comparacoes (Quadrados Latinos)

Análises múltiplas para ANOVA de delineamentos em quadrados latinos

comparacoes_lsqd(
  dados,
  x,
  y,
  linha,
  coluna,
  bloco = NULL,
  alpha = 0.05,
  group = FALSE
)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, linhas, colunas e, caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

linha
String com o nome da coluna dos linha

coluna
String com o nome da coluna dos coluna

bloco
String com o nome da coluna dos blocos, podendo ser as replicas, ou outro tratamento no caso de quadrados greco-latinos

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

Value

Objeto do tipo comparacao_lsqd, que contém:

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

dados Dados alimentados a função

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

Examples

quadrados_latinos(4)$dados %>%
  comparacoes_lsqd("tratamento", "resultado", "linha", "coluna")

#> $variaveis
#> $variaveis$tratamento
#> [1] "tratamento"
#> 
#> $variaveis$linha
#> [1] "linha"
#> 
#> $variaveis$coluna
#> [1] "coluna"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ tratamento + linha + coluna
#> <environment: 0x5618e15966f8>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 tratamento     linha    coluna Residuals
#> Sum of Squares   11.168576  4.483627  0.610603  1.015550
#> Deg. of Freedom          3         3         3         6
#> 
#> Residual standard error: 0.4114102
#> Estimated effects may be unbalanced
#> 
#> $modelos$fisher
#> $statistics
#>     MSerror Df        Mean        CV  t.value       LSD
#>   0.1692583  6 -0.02148927 -1914.491 2.446912 0.7118334
#> 
#> $parameters
#>         test p.ajusted     name.t ntr alpha
#>   Fisher-LSD      none tratamento   4  0.05
#> 
#> $means
#>    resultado       std r        LCL        UCL        Min        Max        Q25
#> 1  1.1046678 0.4540136 4  0.6013256  1.6080100  0.4414282  1.4558414  1.0131829
#> 2  0.3028931 1.1813451 4 -0.2004491  0.8062354 -1.1936412  1.3829109 -0.3568254
#> 3 -0.3185898 0.4758550 4 -0.8219320  0.1847525 -0.9028149  0.1289229 -0.6033984
#> 4 -1.1749283 0.4565640 4 -1.6782705 -0.6715861 -1.6865047 -0.7517233 -1.4950793
#>          Q50         Q75
#> 1  1.2607008  1.35218562
#> 2  0.5111515  1.17087002
#> 3 -0.2502335  0.03457518
#> 4 -1.1307425 -0.81059153
#> 
#> $comparison
#>       difference pvalue signif.         LCL      UCL
#> 1 - 2  0.8017746 0.0330       *  0.08994127 1.513608
#> 1 - 3  1.4232575 0.0027      **  0.71142417 2.135091
#> 1 - 4  2.2795961 0.0002     ***  1.56776269 2.991429
#> 2 - 3  0.6214829 0.0765       . -0.09035047 1.333316
#> 2 - 4  1.4778214 0.0023      **  0.76598805 2.189655
#> 3 - 4  0.8563385 0.0258       *  0.14450515 1.568172
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$bonferroni
#> $statistics
#>     MSerror Df        Mean        CV  t.value      MSD
#>   0.1692583  6 -0.02148927 -1914.491 3.862991 1.123786
#> 
#> $parameters
#>         test  p.ajusted     name.t ntr alpha
#>   Fisher-LSD bonferroni tratamento   4  0.05
#> 
#> $means
#>    resultado       std r        LCL        UCL        Min        Max        Q25
#> 1  1.1046678 0.4540136 4  0.6013256  1.6080100  0.4414282  1.4558414  1.0131829
#> 2  0.3028931 1.1813451 4 -0.2004491  0.8062354 -1.1936412  1.3829109 -0.3568254
#> 3 -0.3185898 0.4758550 4 -0.8219320  0.1847525 -0.9028149  0.1289229 -0.6033984
#> 4 -1.1749283 0.4565640 4 -1.6782705 -0.6715861 -1.6865047 -0.7517233 -1.4950793
#>          Q50         Q75
#> 1  1.2607008  1.35218562
#> 2  0.5111515  1.17087002
#> 3 -0.2502335  0.03457518
#> 4 -1.1307425 -0.81059153
#> 
#> $comparison
#>       difference pvalue signif.        LCL      UCL
#> 1 - 2  0.8017746 0.1981         -0.3220115 1.925561
#> 1 - 3  1.4232575 0.0164       *  0.2994714 2.547044
#> 1 - 4  2.2795961 0.0014      **  1.1558099 3.403382
#> 2 - 3  0.6214829 0.4592         -0.5023032 1.745269
#> 2 - 4  1.4778214 0.0136       *  0.3540353 2.601608
#> 3 - 4  0.8563385 0.1550         -0.2674476 1.980125
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $statistics
#>     MSerror Df        Mean        CV     MSD
#>   0.1692583  6 -0.02148927 -1914.491 1.00705
#> 
#> $parameters
#>    test     name.t ntr StudentizedRange alpha
#>   Tukey tratamento   4         4.895599  0.05
#> 
#> $means
#>    resultado       std r        Min        Max        Q25        Q50
#> 1  1.1046678 0.4540136 4  0.4414282  1.4558414  1.0131829  1.2607008
#> 2  0.3028931 1.1813451 4 -1.1936412  1.3829109 -0.3568254  0.5111515
#> 3 -0.3185898 0.4758550 4 -0.9028149  0.1289229 -0.6033984 -0.2502335
#> 4 -1.1749283 0.4565640 4 -1.6865047 -0.7517233 -1.4950793 -1.1307425
#>           Q75
#> 1  1.35218562
#> 2  1.17087002
#> 3  0.03457518
#> 4 -0.81059153
#> 
#> $comparison
#>       difference pvalue signif.        LCL      UCL
#> 1 - 2  0.8017746 0.1153         -0.2052750 1.808824
#> 1 - 3  1.4232575 0.0108       *  0.4162079 2.430307
#> 1 - 4  2.2795961 0.0009     ***  1.2725464 3.286646
#> 2 - 3  0.6214829 0.2429         -0.3855667 1.628533
#> 2 - 4  1.4778214 0.0090      **  0.4707718 2.484871
#> 3 - 4  0.8563385 0.0920       . -0.1507111 1.863388
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $statistics
#>     MSerror Df        Mean        CV
#>   0.1692583  6 -0.02148927 -1914.491
#> 
#> $parameters
#>     test     name.t ntr alpha
#>   Duncan tratamento   4  0.05
#> 
#> $duncan
#>      Table CriticalRange
#> 2 3.460456     0.7118333
#> 3 3.586498     0.7377608
#> 4 3.648934     0.7506043
#> 
#> $means
#>    resultado       std r        Min        Max        Q25        Q50
#> 1  1.1046678 0.4540136 4  0.4414282  1.4558414  1.0131829  1.2607008
#> 2  0.3028931 1.1813451 4 -1.1936412  1.3829109 -0.3568254  0.5111515
#> 3 -0.3185898 0.4758550 4 -0.9028149  0.1289229 -0.6033984 -0.2502335
#> 4 -1.1749283 0.4565640 4 -1.6865047 -0.7517233 -1.4950793 -1.1307425
#>           Q75
#> 1  1.35218562
#> 2  1.17087002
#> 3  0.03457518
#> 4 -0.81059153
#> 
#> $comparison
#>       difference pvalue signif.         LCL      UCL
#> 1 - 2  0.8017746 0.0330       *  0.08994132 1.513608
#> 1 - 3  1.4232575 0.0033      **  0.68549673 2.161018
#> 1 - 4  2.2795961 0.0003     ***  1.52899180 3.030200
#> 2 - 3  0.6214829 0.0765       . -0.09035042 1.333316
#> 2 - 4  1.4778214 0.0027      **  0.74006061 2.215582
#> 3 - 4  0.8563385 0.0258       *  0.14450520 1.568172
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 16 x 5
#>    resultado tratamento coluna linha replica
#>        <dbl> <fct>      <fct>  <fct>   <int>
#>  1  -1.69    4          1      1           1
#>  2  -0.903   3          1      2           1
#>  3   1.32    1          1      3           1
#>  4   1.10    2          1      4           1
#>  5   1.20    1          2      1           1
#>  6  -1.43    4          2      2           1
#>  7   1.38    2          2      3           1
#>  8   0.00313 3          2      4           1
#>  9  -0.0779  2          3      1           1
#> 10   0.441   1          3      2           1
#> 11   0.129   3          3      3           1
#> 12  -0.830   4          3      4           1
#> 13  -0.504   3          4      1           1
#> 14  -1.19    2          4      2           1
#> 15  -0.752   4          4      3           1
#> 16   1.46    1          4      4           1
#> 
#> $fisher
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2           0.8017746      0.08994127        1.513608  0.0330
#> 2      1 - 3           1.4232575      0.71142417        2.135091  0.0027
#> 3      1 - 4           2.2795961      1.56776269        2.991429  0.0002
#> 4      2 - 3           0.6214829     -0.09035047        1.333316  0.0765
#> 5      2 - 4           1.4778214      0.76598805        2.189655  0.0023
#> 6      3 - 4           0.8563385      0.14450515        1.568172  0.0258
#> 
#> $bonferroni
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2           0.8017746      -0.3220115        1.925561  0.1981
#> 2      1 - 3           1.4232575       0.2994714        2.547044  0.0164
#> 3      1 - 4           2.2795961       1.1558099        3.403382  0.0014
#> 4      2 - 3           0.6214829      -0.5023032        1.745269  0.4592
#> 5      2 - 4           1.4778214       0.3540353        2.601608  0.0136
#> 6      3 - 4           0.8563385      -0.2674476        1.980125  0.1550
#> 
#> $tukey
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2           0.8017746      -0.2052750        1.808824  0.1153
#> 2      1 - 3           1.4232575       0.4162079        2.430307  0.0108
#> 3      1 - 4           2.2795961       1.2725464        3.286646  0.0009
#> 4      2 - 3           0.6214829      -0.3855667        1.628533  0.2429
#> 5      2 - 4           1.4778214       0.4707718        2.484871  0.0090
#> 6      3 - 4           0.8563385      -0.1507111        1.863388  0.0920
#> 
#> $duncan
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2           0.8017746      0.08994132        1.513608  0.0330
#> 2      1 - 3           1.4232575      0.68549673        2.161018  0.0033
#> 3      1 - 4           2.2795961      1.52899180        3.030200  0.0003
#> 4      2 - 3           0.6214829     -0.09035042        1.333316  0.0765
#> 5      2 - 4           1.4778214      0.74006061        2.215582  0.0027
#> 6      3 - 4           0.8563385      0.14450520        1.568172  0.0258
#> 
#> attr(,"class")
#> [1] "comparacao_lsqd"