Comparacoes (Quadrados Latinos)

Análises múltiplas para ANOVA de delineamentos em quadrados latinos

comparacoes_lsqd(dados, x, y, linha, coluna, alpha = 0.05, group = FALSE)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

bloco
String com o nome da coluna dos blocos, manter NULL caso não hajam blocos

Value

Objeto do tipo comparacao_lsqd, que contém:

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

dados Dados alimentados a função

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

Examples

quadrados_latinos(4)$dados %>%
  comparacoes_lsqd("tratamento", "resultado", "linha", "coluna")

#> $variaveis
#> $variaveis$tratamento
#> [1] "tratamento"
#> 
#> $variaveis$linha
#> [1] "linha"
#> 
#> $variaveis$coluna
#> [1] "coluna"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ tratamento + linha + coluna
#> <environment: 0x559601f062a0>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 tratamento    linha   coluna Residuals
#> Sum of Squares    1.578336 1.525516 2.881114  6.850935
#> Deg. of Freedom          3        3        3         6
#> 
#> Residual standard error: 1.068561
#> Estimated effects may be unbalanced
#> 
#> $modelos$fisher
#> $statistics
#>    MSerror Df       Mean       CV  t.value      LSD
#>   1.141823  6 0.08231753 1298.096 2.446912 1.848854
#> 
#> $parameters
#>         test p.ajusted     name.t ntr alpha
#>   Fisher-LSD      none tratamento   4  0.05
#> 
#> $means
#>    resultado      std r        LCL       UCL        Min       Max        Q25
#> 1 -0.4492274 1.317980 4 -1.7565647 0.8581098 -2.2181411 0.9721363 -0.7815650
#> 2  0.2577184 0.396452 4 -1.0496188 1.5650557 -0.2753020 0.6752700  0.1253514
#> 3  0.3548858 1.108942 4 -0.9524515 1.6622230 -0.6098583 1.4172813 -0.5989763
#> 4  0.1658933 0.792795 4 -1.1414439 1.4732306 -0.8671011 1.0439389 -0.1342075
#>          Q50        Q75
#> 1 -0.2754524 0.05688515
#> 2  0.3154529 0.44781993
#> 3  0.3060601 1.25992218
#> 4  0.2433677 0.54346854
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.70694586 0.3856         -2.555800 1.141908
#> 1 - 3 -0.80411321 0.3282         -2.652967 1.044741
#> 1 - 4 -0.61512076 0.4467         -2.463975 1.233733
#> 2 - 3 -0.09716735 0.9019         -1.946021 1.751687
#> 2 - 4  0.09182510 0.9072         -1.757029 1.940679
#> 3 - 4  0.18899246 0.8108         -1.659862 2.037847
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$bonferroni
#> $statistics
#>    MSerror Df       Mean       CV  t.value      MSD
#>   1.141823  6 0.08231753 1298.096 3.862991 2.918824
#> 
#> $parameters
#>         test  p.ajusted     name.t ntr alpha
#>   Fisher-LSD bonferroni tratamento   4  0.05
#> 
#> $means
#>    resultado      std r        LCL       UCL        Min       Max        Q25
#> 1 -0.4492274 1.317980 4 -1.7565647 0.8581098 -2.2181411 0.9721363 -0.7815650
#> 2  0.2577184 0.396452 4 -1.0496188 1.5650557 -0.2753020 0.6752700  0.1253514
#> 3  0.3548858 1.108942 4 -0.9524515 1.6622230 -0.6098583 1.4172813 -0.5989763
#> 4  0.1658933 0.792795 4 -1.1414439 1.4732306 -0.8671011 1.0439389 -0.1342075
#>          Q50        Q75
#> 1 -0.2754524 0.05688515
#> 2  0.3154529 0.44781993
#> 3  0.3060601 1.25992218
#> 4  0.2433677 0.54346854
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.70694586      1         -3.625770 2.211878
#> 1 - 3 -0.80411321      1         -3.722938 2.114711
#> 1 - 4 -0.61512076      1         -3.533945 2.303704
#> 2 - 3 -0.09716735      1         -3.015992 2.821657
#> 2 - 4  0.09182510      1         -2.826999 3.010649
#> 3 - 4  0.18899246      1         -2.729832 3.107817
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $statistics
#>    MSerror Df       Mean       CV      MSD
#>   1.141823  6 0.08231753 1298.096 2.615623
#> 
#> $parameters
#>    test     name.t ntr StudentizedRange alpha
#>   Tukey tratamento   4         4.895599  0.05
#> 
#> $means
#>    resultado      std r        Min       Max        Q25        Q50        Q75
#> 1 -0.4492274 1.317980 4 -2.2181411 0.9721363 -0.7815650 -0.2754524 0.05688515
#> 2  0.2577184 0.396452 4 -0.2753020 0.6752700  0.1253514  0.3154529 0.44781993
#> 3  0.3548858 1.108942 4 -0.6098583 1.4172813 -0.5989763  0.3060601 1.25992218
#> 4  0.1658933 0.792795 4 -0.8671011 1.0439389 -0.1342075  0.2433677 0.54346854
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.70694586 0.7882         -3.322569 1.908677
#> 1 - 3 -0.80411321 0.7217         -3.419736 1.811510
#> 1 - 4 -0.61512076 0.8460         -3.230744 2.000502
#> 2 - 3 -0.09716735 0.9992         -2.712790 2.518456
#> 2 - 4  0.09182510 0.9993         -2.523798 2.707448
#> 3 - 4  0.18899246 0.9939         -2.426631 2.804616
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $statistics
#>    MSerror Df       Mean       CV
#>   1.141823  6 0.08231753 1298.096
#> 
#> $parameters
#>     test     name.t ntr alpha
#>   Duncan tratamento   4  0.05
#> 
#> $duncan
#>      Table CriticalRange
#> 2 3.460456      1.848854
#> 3 3.586498      1.916196
#> 4 3.648934      1.949554
#> 
#> $means
#>    resultado      std r        Min       Max        Q25        Q50        Q75
#> 1 -0.4492274 1.317980 4 -2.2181411 0.9721363 -0.7815650 -0.2754524 0.05688515
#> 2  0.2577184 0.396452 4 -0.2753020 0.6752700  0.1253514  0.3154529 0.44781993
#> 3  0.3548858 1.108942 4 -0.6098583 1.4172813 -0.5989763  0.3060601 1.25992218
#> 4  0.1658933 0.792795 4 -0.8671011 1.0439389 -0.1342075  0.2433677 0.54346854
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.70694586 0.3998         -2.623142 1.209250
#> 1 - 3 -0.80411321 0.3471         -2.753667 1.145441
#> 1 - 4 -0.61512076 0.4467         -2.463975 1.233733
#> 2 - 3 -0.09716735 0.9019         -1.946021 1.751687
#> 2 - 4  0.09182510 0.9072         -1.757029 1.940679
#> 3 - 4  0.18899246 0.8164         -1.727203 2.105188
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 16 x 5
#>    resultado tratamento coluna linha replica
#>        <dbl> <fct>      <fct>  <fct>   <int>
#>  1    -0.248 1          1      1           1
#>  2     0.675 2          1      2           1
#>  3    -0.595 3          1      3           1
#>  4     0.110 4          1      4           1
#>  5     0.372 2          2      1           1
#>  6    -0.610 3          2      2           1
#>  7     1.04  4          2      3           1
#>  8    -0.303 1          2      4           1
#>  9     1.42  3          3      1           1
#> 10    -0.867 4          3      2           1
#> 11    -2.22  1          3      3           1
#> 12    -0.275 2          3      4           1
#> 13     0.377 4          4      1           1
#> 14     0.972 1          4      2           1
#> 15     0.259 2          4      3           1
#> 16     1.21  3          4      4           1
#> 
#> $fisher
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.70694586       -2.555800        1.141908  0.3856
#> 2      1 - 3         -0.80411321       -2.652967        1.044741  0.3282
#> 3      1 - 4         -0.61512076       -2.463975        1.233733  0.4467
#> 4      2 - 3         -0.09716735       -1.946021        1.751687  0.9019
#> 5      2 - 4          0.09182510       -1.757029        1.940679  0.9072
#> 6      3 - 4          0.18899246       -1.659862        2.037847  0.8108
#> 
#> $bonferroni
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.70694586       -3.625770        2.211878       1
#> 2      1 - 3         -0.80411321       -3.722938        2.114711       1
#> 3      1 - 4         -0.61512076       -3.533945        2.303704       1
#> 4      2 - 3         -0.09716735       -3.015992        2.821657       1
#> 5      2 - 4          0.09182510       -2.826999        3.010649       1
#> 6      3 - 4          0.18899246       -2.729832        3.107817       1
#> 
#> $tukey
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.70694586       -3.322569        1.908677  0.7882
#> 2      1 - 3         -0.80411321       -3.419736        1.811510  0.7217
#> 3      1 - 4         -0.61512076       -3.230744        2.000502  0.8460
#> 4      2 - 3         -0.09716735       -2.712790        2.518456  0.9992
#> 5      2 - 4          0.09182510       -2.523798        2.707448  0.9993
#> 6      3 - 4          0.18899246       -2.426631        2.804616  0.9939
#> 
#> $duncan
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.70694586       -2.623142        1.209250  0.3998
#> 2      1 - 3         -0.80411321       -2.753667        1.145441  0.3471
#> 3      1 - 4         -0.61512076       -2.463975        1.233733  0.4467
#> 4      2 - 3         -0.09716735       -1.946021        1.751687  0.9019
#> 5      2 - 4          0.09182510       -1.757029        1.940679  0.9072
#> 6      3 - 4          0.18899246       -1.727203        2.105188  0.8164
#> 
#> attr(,"class")
#> [1] "comparacao_lsqd"