Comparacoes (Quadrados Latinos)

Análises múltiplas para ANOVA de delineamentos em quadrados latinos

comparacoes_lsqd(
  dados,
  x,
  y,
  linha,
  coluna,
  bloco = NULL,
  alpha = 0.05,
  group = FALSE
)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, linhas, colunas e, caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

linha
String com o nome da coluna dos linha

coluna
String com o nome da coluna dos coluna

bloco
String com o nome da coluna dos blocos, podendo ser as replicas, ou outro tratamento no caso de quadrados greco-latinos

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

Value

Objeto do tipo comparacao_lsqd, que contém:

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

dados Dados alimentados a função

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

Examples

quadrados_latinos(4)$dados %>%
  comparacoes_lsqd("tratamento", "resultado", "linha", "coluna")

#> $variaveis
#> $variaveis$tratamento
#> [1] "tratamento"
#> 
#> $variaveis$linha
#> [1] "linha"
#> 
#> $variaveis$coluna
#> [1] "coluna"
#> 
#> $variaveis$resultado
#> [1] "resultado"
#> 
#> 
#> $formula
#> resultado ~ tratamento + linha + coluna
#> <environment: 0x55835c750ae0>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 tratamento     linha    coluna Residuals
#> Sum of Squares    2.194060  1.087180  0.603553 17.391115
#> Deg. of Freedom          3         3         3         6
#> 
#> Residual standard error: 1.702504
#> Estimated effects may be unbalanced
#> 
#> $modelos$fisher
#> $statistics
#>    MSerror Df       Mean        CV  t.value     LSD
#>   2.898519  6 -0.2059605 -826.6168 2.446912 2.94572
#> 
#> $parameters
#>         test p.ajusted     name.t ntr alpha
#>   Fisher-LSD      none tratamento   4  0.05
#> 
#> $means
#>    resultado       std r       LCL      UCL        Min       Max        Q25
#> 1 -0.1156501 0.8313976 4 -2.198588 1.967288 -1.2741013 0.5692981 -0.4333688
#> 2  0.2143972 1.5260584 4 -1.868541 2.297336 -0.8209799 2.4633691 -0.6834682
#> 3 -0.1197775 1.1639466 4 -2.202716 1.963161 -1.4882559 1.0545030 -0.8615702
#> 4 -0.8028114 1.4091729 4 -2.885750 1.280127 -1.9328933 1.2487550 -1.5455282
#>          Q50        Q75
#> 1  0.1211014  0.4388202
#> 2 -0.3924002  0.5054653
#> 3 -0.0226786  0.7191141
#> 4 -1.2635537 -0.5208369
#> 
#> $comparison
#>         difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.330047297 0.7932         -3.275767 2.615672
#> 1 - 3  0.004127465 0.9974         -2.941592 2.949847
#> 1 - 4  0.687161356 0.5889         -2.258558 3.632881
#> 2 - 3  0.334174762 0.7906         -2.611545 3.279894
#> 2 - 4  1.017208654 0.4305         -1.928511 3.962928
#> 3 - 4  0.683033891 0.5910         -2.262686 3.628754
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$bonferroni
#> $statistics
#>    MSerror Df       Mean        CV  t.value      MSD
#>   2.898519  6 -0.2059605 -826.6168 3.862991 4.650469
#> 
#> $parameters
#>         test  p.ajusted     name.t ntr alpha
#>   Fisher-LSD bonferroni tratamento   4  0.05
#> 
#> $means
#>    resultado       std r       LCL      UCL        Min       Max        Q25
#> 1 -0.1156501 0.8313976 4 -2.198588 1.967288 -1.2741013 0.5692981 -0.4333688
#> 2  0.2143972 1.5260584 4 -1.868541 2.297336 -0.8209799 2.4633691 -0.6834682
#> 3 -0.1197775 1.1639466 4 -2.202716 1.963161 -1.4882559 1.0545030 -0.8615702
#> 4 -0.8028114 1.4091729 4 -2.885750 1.280127 -1.9328933 1.2487550 -1.5455282
#>          Q50        Q75
#> 1  0.1211014  0.4388202
#> 2 -0.3924002  0.5054653
#> 3 -0.0226786  0.7191141
#> 4 -1.2635537 -0.5208369
#> 
#> $comparison
#>         difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.330047297      1         -4.980516 4.320422
#> 1 - 3  0.004127465      1         -4.646341 4.654596
#> 1 - 4  0.687161356      1         -3.963308 5.337630
#> 2 - 3  0.334174762      1         -4.316294 4.984644
#> 2 - 4  1.017208654      1         -3.633260 5.667678
#> 3 - 4  0.683033891      1         -3.967435 5.333503
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $statistics
#>    MSerror Df       Mean        CV      MSD
#>   2.898519  6 -0.2059605 -826.6168 4.167388
#> 
#> $parameters
#>    test     name.t ntr StudentizedRange alpha
#>   Tukey tratamento   4         4.895599  0.05
#> 
#> $means
#>    resultado       std r        Min       Max        Q25        Q50        Q75
#> 1 -0.1156501 0.8313976 4 -1.2741013 0.5692981 -0.4333688  0.1211014  0.4388202
#> 2  0.2143972 1.5260584 4 -0.8209799 2.4633691 -0.6834682 -0.3924002  0.5054653
#> 3 -0.1197775 1.1639466 4 -1.4882559 1.0545030 -0.8615702 -0.0226786  0.7191141
#> 4 -0.8028114 1.4091729 4 -1.9328933 1.2487550 -1.5455282 -1.2635537 -0.5208369
#> 
#> $comparison
#>         difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.330047297 0.9921         -4.497435 3.837341
#> 1 - 3  0.004127465 1.0000         -4.163261 4.171516
#> 1 - 4  0.687161356 0.9373         -3.480227 4.854549
#> 2 - 3  0.334174762 0.9918         -3.833213 4.501563
#> 2 - 4  1.017208654 0.8319         -3.150179 5.184597
#> 3 - 4  0.683033891 0.9383         -3.484354 4.850422
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $statistics
#>    MSerror Df       Mean        CV
#>   2.898519  6 -0.2059605 -826.6168
#> 
#> $parameters
#>     test     name.t ntr alpha
#>   Duncan tratamento   4  0.05
#> 
#> $duncan
#>      Table CriticalRange
#> 2 3.460456      2.945719
#> 3 3.586498      3.053013
#> 4 3.648934      3.106162
#> 
#> $means
#>    resultado       std r        Min       Max        Q25        Q50        Q75
#> 1 -0.1156501 0.8313976 4 -1.2741013 0.5692981 -0.4333688  0.1211014  0.4388202
#> 2  0.2143972 1.5260584 4 -0.8209799 2.4633691 -0.6834682 -0.3924002  0.5054653
#> 3 -0.1197775 1.1639466 4 -1.4882559 1.0545030 -0.8615702 -0.0226786  0.7191141
#> 4 -0.8028114 1.4091729 4 -1.9328933 1.2487550 -1.5455282 -1.2635537 -0.5208369
#> 
#> $comparison
#>         difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.330047297 0.7932         -3.275767 2.615672
#> 1 - 3  0.004127465 0.9974         -2.941592 2.949847
#> 1 - 4  0.687161356 0.6000         -2.365852 3.740174
#> 2 - 3  0.334174762 0.7968         -2.718838 3.387188
#> 2 - 4  1.017208654 0.4481         -2.088953 4.123371
#> 3 - 4  0.683033891 0.5910         -2.262686 3.628753
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 16 x 5
#>    resultado tratamento coluna linha replica
#>        <dbl> <fct>      <fct>  <fct>   <int>
#>  1    -1.49  3          1      1           1
#>  2     2.46  2          1      2           1
#>  3     0.569 1          1      3           1
#>  4    -1.11  4          1      4           1
#>  5     1.25  4          2      1           1
#>  6    -0.653 3          2      2           1
#>  7    -0.147 2          2      3           1
#>  8    -1.27  1          2      4           1
#>  9     0.395 1          3      1           1
#> 10    -1.93  4          3      2           1
#> 11     1.05  3          3      3           1
#> 12    -0.821 2          3      4           1
#> 13    -0.638 2          4      1           1
#> 14    -0.153 1          4      2           1
#> 15    -1.42  4          4      3           1
#> 16     0.607 3          4      4           1
#> 
#> $fisher
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2        -0.330047297       -3.275767        2.615672  0.7932
#> 2      1 - 3         0.004127465       -2.941592        2.949847  0.9974
#> 3      1 - 4         0.687161356       -2.258558        3.632881  0.5889
#> 4      2 - 3         0.334174762       -2.611545        3.279894  0.7906
#> 5      2 - 4         1.017208654       -1.928511        3.962928  0.4305
#> 6      3 - 4         0.683033891       -2.262686        3.628754  0.5910
#> 
#> $bonferroni
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2        -0.330047297       -4.980516        4.320422       1
#> 2      1 - 3         0.004127465       -4.646341        4.654596       1
#> 3      1 - 4         0.687161356       -3.963308        5.337630       1
#> 4      2 - 3         0.334174762       -4.316294        4.984644       1
#> 5      2 - 4         1.017208654       -3.633260        5.667678       1
#> 6      3 - 4         0.683033891       -3.967435        5.333503       1
#> 
#> $tukey
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2        -0.330047297       -4.497435        3.837341  0.9921
#> 2      1 - 3         0.004127465       -4.163261        4.171516  1.0000
#> 3      1 - 4         0.687161356       -3.480227        4.854549  0.9373
#> 4      2 - 3         0.334174762       -3.833213        4.501563  0.9918
#> 5      2 - 4         1.017208654       -3.150179        5.184597  0.8319
#> 6      3 - 4         0.683033891       -3.484354        4.850422  0.9383
#> 
#> $duncan
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2        -0.330047297       -3.275767        2.615672  0.7932
#> 2      1 - 3         0.004127465       -2.941592        2.949847  0.9974
#> 3      1 - 4         0.687161356       -2.365852        3.740174  0.6000
#> 4      2 - 3         0.334174762       -2.718838        3.387188  0.7968
#> 5      2 - 4         1.017208654       -2.088953        4.123371  0.4481
#> 6      3 - 4         0.683033891       -2.262686        3.628753  0.5910
#> 
#> attr(,"class")
#> [1] "comparacao_lsqd"