Comparacoes (Quadrados Latinos)

Análises múltiplas para ANOVA de delineamentos em quadrados latinos

comparacoes_lsqd(dados, x, y, linha, coluna, alpha = 0.05, group = FALSE)

Arguments

dados
Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos

x
String com o nome da coluna dos tratamentos

y
String com o nome da coluna dos resultados

alpha
Nivel de significancia a ser utilizado, padrao 0.05

group
Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos

bloco
String com o nome da coluna dos blocos, manter NULL caso não hajam blocos

Value

Objeto do tipo comparacao_lsqd, que contém:

variaveis Lista com os nomes das variaveis

formula Objeto formula usado para criar o modelo

modelos Lista com todos os modelos criados

dados Dados alimentados a função

fisher Dados das comparações de fisher

bonferroni Dados das comparações de bonferroni

tukey Dados das comparações de tukey

duncan Dados das comparações de duncan

Examples

quadrados_latinos(tibble(y = rnorm(16)), 1:4) %>%
  comparacoes_lsqd("tratamento", "y", "linha", "coluna")

#> $variaveis
#> $variaveis$tratamento
#> [1] "tratamento"
#> 
#> $variaveis$linha
#> [1] "linha"
#> 
#> $variaveis$coluna
#> [1] "coluna"
#> 
#> $variaveis$resultado
#> [1] "y"
#> 
#> 
#> $formula
#> y ~ tratamento + linha + coluna
#> <environment: 0x563d2e811000>
#> 
#> $modelos
#> $modelos$Modelo
#> Call:
#>    aov(formula = formula_, data = dados)
#> 
#> Terms:
#>                 tratamento    linha   coluna Residuals
#> Sum of Squares    0.235990 5.054071 3.613504  4.994708
#> Deg. of Freedom          3        3        3         6
#> 
#> Residual standard error: 0.9123877
#> Estimated effects may be unbalanced
#> 
#> $modelos$fisher
#> $statistics
#>     MSerror Df      Mean       CV  t.value      LSD
#>   0.8324514  6 0.1855902 491.6142 2.446912 1.578639
#> 
#> $parameters
#>         test p.ajusted     name.t ntr alpha
#>   Fisher-LSD      none tratamento   4  0.05
#> 
#> $means
#>            y       std r        LCL      UCL        Min       Max        Q25
#> 1 0.03945468 0.8034924 4 -1.0768115 1.155721 -0.7020820 0.9763418 -0.5904491
#> 2 0.20107459 1.2886263 4 -0.9151916 1.317341 -1.0405791 1.4051089 -0.8432721
#> 3 0.13077553 1.2578007 4 -0.9854907 1.247042 -1.5665145 1.1465005 -0.4492419
#> 4 0.37105592 0.8160111 4 -0.7452103 1.487322 -0.7954609 0.9733158  0.1112177
#>           Q50       Q75
#> 1 -0.05822054 0.5716833
#> 2  0.21988430 1.2642310
#> 3  0.47155805 1.0515754
#> 4  0.65318443 0.9130226
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.16161991 0.8105         -1.740259 1.417019
#> 1 - 3 -0.09132085 0.8921         -1.669960 1.487318
#> 1 - 4 -0.33160124 0.6256         -1.910240 1.247038
#> 2 - 3  0.07029906 0.9168         -1.508340 1.648938
#> 2 - 4 -0.16998133 0.8010         -1.748620 1.408657
#> 3 - 4 -0.24028039 0.7224         -1.818919 1.338358
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$bonferroni
#> $statistics
#>     MSerror Df      Mean       CV  t.value     MSD
#>   0.8324514  6 0.1855902 491.6142 3.862991 2.49223
#> 
#> $parameters
#>         test  p.ajusted     name.t ntr alpha
#>   Fisher-LSD bonferroni tratamento   4  0.05
#> 
#> $means
#>            y       std r        LCL      UCL        Min       Max        Q25
#> 1 0.03945468 0.8034924 4 -1.0768115 1.155721 -0.7020820 0.9763418 -0.5904491
#> 2 0.20107459 1.2886263 4 -0.9151916 1.317341 -1.0405791 1.4051089 -0.8432721
#> 3 0.13077553 1.2578007 4 -0.9854907 1.247042 -1.5665145 1.1465005 -0.4492419
#> 4 0.37105592 0.8160111 4 -0.7452103 1.487322 -0.7954609 0.9733158  0.1112177
#>           Q50       Q75
#> 1 -0.05822054 0.5716833
#> 2  0.21988430 1.2642310
#> 3  0.47155805 1.0515754
#> 4  0.65318443 0.9130226
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.16161991      1         -2.653850 2.330610
#> 1 - 3 -0.09132085      1         -2.583551 2.400909
#> 1 - 4 -0.33160124      1         -2.823831 2.160629
#> 2 - 3  0.07029906      1         -2.421931 2.562529
#> 2 - 4 -0.16998133      1         -2.662211 2.322249
#> 3 - 4 -0.24028039      1         -2.732510 2.251949
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$tukey
#> $statistics
#>     MSerror Df      Mean       CV      MSD
#>   0.8324514  6 0.1855902 491.6142 2.233342
#> 
#> $parameters
#>    test     name.t ntr StudentizedRange alpha
#>   Tukey tratamento   4         4.895599  0.05
#> 
#> $means
#>            y       std r        Min       Max        Q25         Q50       Q75
#> 1 0.03945468 0.8034924 4 -0.7020820 0.9763418 -0.5904491 -0.05822054 0.5716833
#> 2 0.20107459 1.2886263 4 -1.0405791 1.4051089 -0.8432721  0.21988430 1.2642310
#> 3 0.13077553 1.2578007 4 -1.5665145 1.1465005 -0.4492419  0.47155805 1.0515754
#> 4 0.37105592 0.8160111 4 -0.7954609 0.9733158  0.1112177  0.65318443 0.9130226
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.16161991 0.9939         -2.394962 2.071722
#> 1 - 3 -0.09132085 0.9989         -2.324663 2.142021
#> 1 - 4 -0.33160124 0.9527         -2.564944 1.901741
#> 2 - 3  0.07029906 0.9995         -2.163043 2.303641
#> 2 - 4 -0.16998133 0.9929         -2.403324 2.063361
#> 3 - 4 -0.24028039 0.9808         -2.473623 1.993062
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> $modelos$duncan
#> $statistics
#>     MSerror Df      Mean       CV
#>   0.8324514  6 0.1855902 491.6142
#> 
#> $parameters
#>     test     name.t ntr alpha
#>   Duncan tratamento   4  0.05
#> 
#> $duncan
#>      Table CriticalRange
#> 2 3.460456      1.578639
#> 3 3.586498      1.636138
#> 4 3.648934      1.664621
#> 
#> $means
#>            y       std r        Min       Max        Q25         Q50       Q75
#> 1 0.03945468 0.8034924 4 -0.7020820 0.9763418 -0.5904491 -0.05822054 0.5716833
#> 2 0.20107459 1.2886263 4 -1.0405791 1.4051089 -0.8432721  0.21988430 1.2642310
#> 3 0.13077553 1.2578007 4 -1.5665145 1.1465005 -0.4492419  0.47155805 1.0515754
#> 4 0.37105592 0.8160111 4 -0.7954609 0.9733158  0.1112177  0.65318443 0.9130226
#> 
#> $comparison
#>        difference pvalue signif.       LCL      UCL
#> 1 - 2 -0.16161991 0.8161         -1.797758 1.474518
#> 1 - 3 -0.09132085 0.8921         -1.669960 1.487318
#> 1 - 4 -0.33160124 0.6384         -1.996223 1.333020
#> 2 - 3  0.07029906 0.9168         -1.508340 1.648938
#> 2 - 4 -0.16998133 0.8010         -1.748620 1.408657
#> 3 - 4 -0.24028039 0.7304         -1.876419 1.395858
#> 
#> $groups
#> NULL
#> 
#> attr(,"class")
#> [1] "group"
#> 
#> 
#> $dados
#> # A tibble: 16 x 4
#>          y tratamento coluna linha
#>      <dbl> <fct>      <fct>  <fct>
#>  1 -0.553  1          1      1    
#>  2  1.41   2          1      2    
#>  3 -0.795  4          1      3    
#>  4 -1.57   3          1      4    
#>  5 -1.04   2          2      1    
#>  6  1.02   3          2      2    
#>  7 -0.702  1          2      3    
#>  8  0.973  4          2      4    
#>  9 -0.0768 3          3      1    
#> 10  0.893  4          3      2    
#> 11 -0.778  2          3      3    
#> 12  0.437  1          3      4    
#> 13  0.413  4          4      1    
#> 14  0.976  1          4      2    
#> 15  1.15   3          4      3    
#> 16  1.22   2          4      4    
#> 
#> $fisher
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.16161991       -1.740259        1.417019  0.8105
#> 2      1 - 3         -0.09132085       -1.669960        1.487318  0.8921
#> 3      1 - 4         -0.33160124       -1.910240        1.247038  0.6256
#> 4      2 - 3          0.07029906       -1.508340        1.648938  0.9168
#> 5      2 - 4         -0.16998133       -1.748620        1.408657  0.8010
#> 6      3 - 4         -0.24028039       -1.818919        1.338358  0.7224
#> 
#> $bonferroni
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.16161991       -2.653850        2.330610       1
#> 2      1 - 3         -0.09132085       -2.583551        2.400909       1
#> 3      1 - 4         -0.33160124       -2.823831        2.160629       1
#> 4      2 - 3          0.07029906       -2.421931        2.562529       1
#> 5      2 - 4         -0.16998133       -2.662211        2.322249       1
#> 6      3 - 4         -0.24028039       -2.732510        2.251949       1
#> 
#> $tukey
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.16161991       -2.394962        2.071722  0.9939
#> 2      1 - 3         -0.09132085       -2.324663        2.142021  0.9989
#> 3      1 - 4         -0.33160124       -2.564944        1.901741  0.9527
#> 4      2 - 3          0.07029906       -2.163043        2.303641  0.9995
#> 5      2 - 4         -0.16998133       -2.403324        2.063361  0.9929
#> 6      3 - 4         -0.24028039       -2.473623        1.993062  0.9808
#> 
#> $duncan
#>   Comparação Diferença de Médias Limite Inferior Limite Superior P-valor
#> 1      1 - 2         -0.16161991       -1.797758        1.474518  0.8161
#> 2      1 - 3         -0.09132085       -1.669960        1.487318  0.8921
#> 3      1 - 4         -0.33160124       -1.996223        1.333020  0.6384
#> 4      2 - 3          0.07029906       -1.508340        1.648938  0.9168
#> 5      2 - 4         -0.16998133       -1.748620        1.408657  0.8010
#> 6      3 - 4         -0.24028039       -1.876419        1.395858  0.7304
#> 
#> attr(,"class")
#> [1] "comparacao_lsqd"